Add some structure to the builtin library
