+++ analyzer/src/main/kotlin/com/github/cubuspl42/sigmaLang/analyzer/semantics/builtins/collections/set/SetUnionFunction.kt
@@ -1,57 +1,52 @@
-package com.github.cubuspl42.sigmaLang.analyzer.semantics.builtins
+package com.github.cubuspl42.sigmaLang.analyzer.semantics.builtins.collections.set
 
-import com.github.cubuspl42.sigmaLang.analyzer.evaluation.values.BoolValue
 import com.github.cubuspl42.sigmaLang.analyzer.evaluation.values.BuiltinGenericFunctionConstructor
 import com.github.cubuspl42.sigmaLang.analyzer.evaluation.values.Identifier
+import com.github.cubuspl42.sigmaLang.analyzer.evaluation.values.IntValue
 import com.github.cubuspl42.sigmaLang.analyzer.evaluation.values.SetValue
 import com.github.cubuspl42.sigmaLang.analyzer.evaluation.values.StrictBuiltinOrderedFunctionConstructor
 import com.github.cubuspl42.sigmaLang.analyzer.evaluation.values.Value
-import com.github.cubuspl42.sigmaLang.analyzer.semantics.expressions.AbstractionConstructor.ArgumentDeclaration
-import com.github.cubuspl42.sigmaLang.analyzer.semantics.types.BoolType
+import com.github.cubuspl42.sigmaLang.analyzer.semantics.types.GenericType.Companion.orderedTraitDeclaration
 import com.github.cubuspl42.sigmaLang.analyzer.semantics.types.OrderedTupleType
 import com.github.cubuspl42.sigmaLang.analyzer.semantics.types.SetType
-import com.github.cubuspl42.sigmaLang.analyzer.semantics.types.TypeType
 import com.github.cubuspl42.sigmaLang.analyzer.semantics.types.TypeVariable
 
-object SetContainsFunction : BuiltinGenericFunctionConstructor() {
-    override val parameterDeclaration = ArgumentDeclaration(
-        declaredType = OrderedTupleType(
-            elements = listOf(
-                OrderedTupleType.Element(
-                    name = Identifier.of("e"),
-                    type = TypeType,
-                ),
-            ),
-        ),
+object SetUnionFunction : BuiltinGenericFunctionConstructor() {
+    override val parameterDeclaration = orderedTraitDeclaration(
+        Identifier.of("e"),
     )
 
     private val eTypeVariable = TypeVariable(
         parameterDeclaration,
-        path = TypeVariable.Path.of(Identifier.of("e")),
+        path = TypeVariable.Path.of(IntValue(value = 0L)),
     )
 
     override val body = object : StrictBuiltinOrderedFunctionConstructor() {
         override val argumentElements = listOf(
             OrderedTupleType.Element(
-                name = Identifier.of("set"),
+                name = Identifier.of("set0"),
                 type = SetType(
                     elementType = eTypeVariable,
                 ),
             ),
             OrderedTupleType.Element(
-                name = Identifier.of("element"),
-                type = eTypeVariable,
+                name = Identifier.of("set1"),
+                type = SetType(
+                    elementType = eTypeVariable,
+                ),
             ),
         )
 
-        override val imageType = BoolType
+        override val imageType = SetType(
+            elementType = eTypeVariable,
+        )
 
         override fun compute(args: List<Value>): Value {
-            val set = (args[0] as SetValue).elements
-            val element = args[1]
+            val set0 = (args[0] as SetValue).elements
+            val set1 = (args[1] as SetValue).elements
 
-            return BoolValue(
-                value = set.contains(element),
+            return SetValue(
+                elements = set0 + set1,
             )
         }
     }
